requirements.txt
tweepy
python-dotenv
requests
python version 11